sales/requirements.txt
asgiref
Django
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
django-extensions